Socket export seem to work after using namespaces correctly. Frontend still broken
