+++ backend/beets_flask/importer/__init__.py
@@ -1,10 +1,9 @@
-from .communicator import WebsocketCommunicator, ImportCommunicator
+from .communicator import ImportCommunicator
 from .session import InteractiveImportSession
 from .states import ImportState
 from . import types
 
 __all__ = [
-    "WebsocketCommunicator",
     "ImportCommunicator",
     "InteractiveImportSession",
     "ImportState",
